+++ lib/src/widget.dart
@@ -7,12 +7,11 @@ class EzLocalizationBuilder extends StatefulWidget {
   final EzLocalizationDelegate delegate;
 
   /// The widget builder.
-  final Widget Function(
-          BuildContext context, EzLocalizationDelegate ezLocalizationDelegate)
-      builder;
+  final Widget Function(BuildContext context, EzLocalizationDelegate ezLocalizationDelegate) builder;
 
   /// Creates a new EzLocalization builder instance.
   const EzLocalizationBuilder({
+    super.key,
     this.delegate = const EzLocalizationDelegate(),
     required this.builder,
   });
@@ -21,8 +20,7 @@ class EzLocalizationBuilder extends StatefulWidget {
   State<StatefulWidget> createState() => EzLocalizationBuilderState();
 
   /// Allows to change the preferred locale (if using the builder).
-  static EzLocalizationBuilderState? of(BuildContext context) =>
-      context.findAncestorStateOfType<EzLocalizationBuilderState>();
+  static EzLocalizationBuilderState? of(BuildContext context) => context.findAncestorStateOfType<EzLocalizationBuilderState>();
 }
 
 /// The EzLocalization builder state.
@@ -37,11 +35,10 @@ class EzLocalizationBuilderState extends State<EzLocalizationBuilder> {
   }
 
   @override
-  Widget build(BuildContext context) =>
-      widget.builder(context, _ezLocalizationDelegate);
+  Widget build(BuildContext context) => widget.builder(context, _ezLocalizationDelegate);
 
   /// Allows to change the preferred locale.
-  void changeLocale(Locale locale) {
+  void changeLocale(Locale? locale) {
     if (mounted) {
       setState(() {
         _ezLocalizationDelegate = EzLocalizationDelegate(